// scripts/types.d.ts
// External contract dependencies
export interface SuberraConfig {
    anchor_market_contract: string;
    aterra_token_contract: string;
}

export interface SuberraContractsConfig {
    jobs_registry_contract: string;
    subwallet_factory_contract: string;
    product_factory_contract: string;
    p2p_contract: string;
    token_stream_contract: string;
    subscription_product_code_id: number;
    p2p_contract_code_id: number;
    subwallet_code_id: number;
}